feat: ✨ Introduce Jira ID if in branch
Prior to this change, the developer would need to ensure the jiraId is
included in every commit message.

This change reduces that toil by detecting the jira id from the branch.
If a jira matches the regex, then the jira id is appended to the message
as (EXP-210) at the end of the message title. If the jira id is not
detected in the message then nothing happens.